Separate Styles from component
