+++ src/Components/UserForm/index.jsx
@@ -5,7 +5,6 @@ import LinearProgress from '@material-ui/core/LinearProgress';
 import getPasswordStrength from 'password-strength-calc';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import Typography from '@material-ui/core/Typography';
-import { makeStyles } from '@material-ui/core/styles';
 import Container from '@material-ui/core/Container';
 import TextField from '@material-ui/core/TextField';
 import Avatar from '@material-ui/core/Avatar';
@@ -17,7 +16,8 @@ import PropTypes from 'prop-types';
 import * as yup from 'yup';
 
 import CustomLink from '../CustomLink';
-import { user } from '../../actions';
+
+import useStyles from './styles';
 
 const schema = yup.object().shape({
   email: yup
@@ -34,31 +34,6 @@ const schema = yup.object().shape({
       },
     ),
 });
-
-const useStyles = makeStyles(theme => ({
-  '@global': {
-    body: {
-      backgroundColor: theme.palette.common.white,
-    },
-  },
-  paper: {
-    marginTop: theme.spacing(8),
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-  },
-  avatar: {
-    margin: theme.spacing(1),
-    backgroundColor: theme.palette.secondary.main,
-  },
-  form: {
-    width: '100%', // Fix IE 11 issue.
-    marginTop: theme.spacing(3),
-  },
-  submit: {
-    margin: theme.spacing(3, 0, 2),
-  },
-}));
 
 function UserForm({ values, onSubmit }) {
   const error = useSelector(state => state.user.error);
